+++ solutions/cookbooks/jmier_mongodb/spec/unit/recipes/yum_repo_spec.rb
@@ -7,23 +7,17 @@
 require 'spec_helper'
 
 describe 'jmier_mongodb::yum_repo' do
-  context 'When all attributes are default, on Ubuntu 20.04' do
+  context 'When all attributes are default, on CentOS 7' do
     # for a complete list of available platforms and versions see:
     # https://github.com/chefspec/fauxhai/blob/master/PLATFORMS.md
-    platform 'ubuntu', '20.04'
+    platform 'centos', '7'
 
     it 'converges successfully' do
       expect { chef_run }.to_not raise_error
     end
-  end
 
-  context 'When all attributes are default, on CentOS 8' do
-    # for a complete list of available platforms and versions see:
-    # https://github.com/chefspec/fauxhai/blob/master/PLATFORMS.md
-    platform 'centos', '8'
-
-    it 'converges successfully' do
-      expect { chef_run }.to_not raise_error
+    it 'creates yum repository mongodb' do
+      expect(chef_run).to create_yum_repository('mongodb')
     end
   end
 end