+++ solutions/cookbooks/jmier_mongodb/test/integration/default/controls/mongodb_test.rb
@@ -1,0 +1,13 @@
+# InSpec test for recipe jmier_mongodb::mongodb
+
+# The InSpec reference, with examples and extensive documentation, can be
+# found at https://docs.chef.io/inspec/resources/
+
+describe package('mongodb-org') do
+  it { should be_installed }
+end
+
+describe service('mongod') do
+  it { should be_enabled }
+  it { should be_running }
+end